pyproject.toml
[project]
name = "medibook"
version = "1.0.0"
description = "Modern Clinic Appointment Booking System"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<5.2",
    "djangorestframework>=3.15.2",
    "psycopg[binary]>=3.2.3",
    "django-htmx>=1.19.0",
    "django-allauth>=65.3.0",
    "django-compressor>=4.5.1",
    "django-libsass>=0.9",
    "django-widget-tweaks>=1.5.0",
    "pillow>=11.0.0",
    "python-decouple>=3.8",
    "django-redis>=5.4.0",
    "redis==4.6.0",
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.8.2",
    "faker>=33.1.0",
    "django-debug-toolbar>=4.4.6",
    "django-extensions>=3.2.3",
    "django-modeltranslation>=0.18.13",
    "django-rosetta>=0.10.1",
    "django-timezone-field>=7.0",
    "python-dateutil>=2.9.0",
    "dj-config-url>=0.1.1",
    "signals>=0.0.2",
    "hiredis==2.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-cov>=6.0.0",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "ipython>=8.31.0",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing"
